move components to better structure
