+++ frontend/src/component/changeRequest/ChangeRequest/Changes/Change/StrategyChange.tsx
